newline v0.0.1 Newline.Web

A module that keeps using definitions for controllers, views and so on.

This can be used in your application as:

use Newline.Web, :controller
use Newline.Web, :view

The definitions below will be executed for every view, controller, etc, so keep them short and clean, focused on imports, uses and aliases.

Do NOT define functions inside the quoted expressions below.

Functions

__using__(which) (macro)

When used, dispatch to the appropriate controller/view/etc.

channel()

controller()

router()

view()