newline v0.0.1 Newline.Web.Router.Helpers

Module with named helpers generated from Newline.Web.Router.

Functions

page_path(conn_or_endpoint, action, path)

page_path(conn_or_endpoint, action, path, params)

page_url(conn_or_endpoint, action, path)

page_url(conn_or_endpoint, action, path, params)

path(data, path)

Generates the path information including any necessary prefix.

session_path(conn_or_endpoint, action)

session_path(conn_or_endpoint, action, params)

session_url(conn_or_endpoint, action)

session_url(conn_or_endpoint, action, params)

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_path(conn_or_endpoint, action)

user_path(conn_or_endpoint, action, params)

user_url(conn_or_endpoint, action)

user_url(conn_or_endpoint, action, params)